% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitit.R
\name{knitit}
\alias{knitit}
\title{knit an Rmd file into a word docx file}
\usage{
knitit(filename = filename, withgenes = FALSE, nerd.report = FALSE,
  myeval = TRUE, format = "portrait", template = NULL, mk.dir = TRUE,
  figwidth = 7.5, figheight = 5, date.it = TRUE)
}
\arguments{
\item{nerd.report}{show code and results by default? (you can override any block yourself)}

\item{myeval}{evaluate code in blocks by default (possibly useful for debugging)}

\item{format}{'portrait' or 'landscape'  if you supply your own docx template this will be ignored}

\item{template}{docx file to be used as template for word document creation}

\item{mk.dir}{make a Reports directory under current wd and place .md and .docx files there?}

\item{figwidth}{default fig.width for graphics}

\item{date.it}{add date string to output docx file name}

\item{filname}{input Rmd file minus the '.Rmd'}

\item{fixheight}{default fig.height for graphics}
}
\value{
name of docx file produced
}
\description{
knit an Rmd file into a word docx file
}

